refactoring a bit more
