+++ stellar/stellar-common/src/main/java/com/caseystella/stellar/common/configuration/ConfigurationsUtils.java
@@ -14,7 +14,7 @@
  */
 package com.caseystella.stellar.common.configuration;
 
-import com.caseystella.sketchy.serialization.JSONUtils;
+import com.caseystella.sketchy.utilities.JSONUtils;
 import com.caseystella.stellar.common.Constants;
 import com.caseystella.stellar.dsl.Context;
 import com.caseystella.stellar.dsl.StellarFunctions;
