+++ stellar/stellar-common/src/main/java/com/caseystella/stellar/common/utils/ConcatMap.java
@@ -16,7 +16,7 @@
 package com.caseystella.stellar.common.utils;
 
 
-import com.caseystella.sketchy.serialization.SerDeUtils;
+import com.caseystella.sketchy.utilities.SerDeUtils;
 import com.esotericsoftware.kryo.Kryo;
 import com.esotericsoftware.kryo.KryoSerializable;
 import com.esotericsoftware.kryo.io.Input;
@@ -35,7 +35,7 @@ import java.util.Set;
  * ConcatMap is a lazy concatenation of a list of Maps. It is lazy in that it does not construct a
  * union of all of the maps, but rather keeps the maps separate. Key/Value resolution is done via a
  * first-wins strategy (i.e. the first map which has a key will be used).
- *
+ * <p>
  * Also, note, that this is an immutable map, so operations which require mutation will have
  * UnsupportedOperationException thrown.
  */
@@ -81,7 +81,6 @@ public class ConcatMap implements Map<String, Object>, Serializable, KryoSeriali
   }
 
   /**
-   *
    * If any maps contains the value, then this will return true.
    *
    * @param value
@@ -128,7 +127,6 @@ public class ConcatMap implements Map<String, Object>, Serializable, KryoSeriali
   }
 
   /**
-   *
    * This is an immutable map and this operation is not supported.
    *
    * @param key
@@ -140,7 +138,6 @@ public class ConcatMap implements Map<String, Object>, Serializable, KryoSeriali
   }
 
   /**
-   *
    * This is an immutable map and this operation is not supported.
    *
    * @param m
@@ -151,7 +148,6 @@ public class ConcatMap implements Map<String, Object>, Serializable, KryoSeriali
   }
 
   /**
-   *
    * This is an immutable map and this operation is not supported.
    */
   @Override
