+++ stellar/stellar-common/src/main/java/com/caseystella/stellar/common/utils/hashing/DefaultHasher.java
@@ -14,7 +14,7 @@
  */
 package com.caseystella.stellar.common.utils.hashing;
 
-import com.caseystella.sketchy.serialization.ConversionUtils;
+import com.caseystella.sketchy.utilities.ConversionUtils;
 import java.io.Serializable;
 import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
