+++ stellar/stellar-common/src/main/java/com/caseystella/stellar/dsl/functions/DataStructureFunctions.java
@@ -14,8 +14,8 @@
  */
 package com.caseystella.stellar.dsl.functions;
 
-import com.caseystella.sketchy.serialization.ConversionUtils;
-import com.caseystella.sketchy.serialization.SerDeUtils;
+import com.caseystella.sketchy.utilities.ConversionUtils;
+import com.caseystella.sketchy.utilities.SerDeUtils;
 import com.caseystella.stellar.common.utils.BloomFilter;
 import com.caseystella.stellar.dsl.BaseStellarFunction;
 import com.caseystella.stellar.dsl.Stellar;
