+++ stellar/stellar-common/src/main/java/com/caseystella/stellar/dsl/functions/ShellFunctions.java
@@ -14,7 +14,7 @@
  */
 package com.caseystella.stellar.dsl.functions;
 
-import com.caseystella.sketchy.serialization.ConversionUtils;
+import com.caseystella.sketchy.utilities.ConversionUtils;
 import com.caseystella.stellar.common.shell.VariableResult;
 import com.caseystella.stellar.common.shell.cli.PausableInput;
 import com.caseystella.stellar.dsl.BaseStellarFunction;
