+++ stellar/stellar-common/src/test/java/com/caseystella/stellar/common/shell/specials/MagicListVariablesTest.java
@@ -20,7 +20,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import com.caseystella.sketchy.serialization.ConversionUtils;
+import com.caseystella.sketchy.utilities.ConversionUtils;
 import com.caseystella.stellar.common.shell.DefaultStellarShellExecutor;
 import com.caseystella.stellar.common.shell.StellarResult;
 import java.util.Arrays;
